maint: Split up big file into smaller parts (#1015)
This is splitting the big source file up into several smaller, more
manageable parts.
